+++ packages/eas-cli/src/commands/worker/deploy.ts
@@ -43,6 +43,7 @@ interface DeployFlags {
   environment?: EnvironmentVariableEnvironment;
   deploymentIdentifier?: string;
   exportDir: string;
+  dryRun: boolean;
 }
 
 interface RawDeployFlags {
@@ -53,6 +54,7 @@ interface RawDeployFlags {
   alias?: string;
   id?: string;
   'export-dir': string;
+  'dry-run': boolean;
 }
 
 interface DeployInProgressParams {
@@ -89,6 +91,10 @@ export default class WorkerDeploy extends EasCommand {
       description: 'Directory where the Expo project was exported.',
       helpValue: 'dir',
       default: 'dist',
+    }),
+    'dry-run': Flags.boolean({
+      description: 'Outputs a tarball of the new deployment instead of uploading it.',
+      default: false,
     }),
     ...EASEnvironmentFlag,
     ...EasNonInteractiveAndJsonFlags,
@@ -265,6 +271,16 @@ export default class WorkerDeploy extends EasCommand {
         })
       );
 
+      if (flags.dryRun) {
+        const DRY_RUN_OUTPUT_PATH = 'deploy.tar.gz';
+        await fs.promises.copyFile(tarPath, DRY_RUN_OUTPUT_PATH);
+        progress.succeed('Saved deploy.tar.gz tarball');
+        if (flags.json) {
+          printJsonOnlyOutput({ tarPath: DRY_RUN_OUTPUT_PATH });
+        }
+        return;
+      }
+
       const uploadUrl = await getSignedDeploymentUrlAsync(graphqlClient, {
         appId: projectId,
         deploymentIdentifier: flags.deploymentIdentifier,
@@ -380,6 +396,7 @@ export default class WorkerDeploy extends EasCommand {
       deploymentIdentifier: flags.id?.trim(),
       exportDir: flags['export-dir'],
       environment: flags['environment'],
+      dryRun: flags['dry-run'],
     };
   }
 }